Add automatic translations with google
Hi,

This is a proposal pr for adding google translate as an automatic translation service.

While the pr needs still a lot of work I wanted to share this draft so we can discuss if it is something good or not for this project.

I have never worked with typescript, I did my best to follow conventions but sorry if I missed out on things.

There are quite some todo's inline that further describe the idea, potential issues and improvements that can be done.

Thank you for this awesome project!